refactor(routes): Relocate `routes/$lang/_protected` to `routes/protected` and update all relevant imports and route ids. (#2174)
This refactor is part of a series aimed at simplifying the application's routing
structure. In this PR, the `routes/$lang/_protected` folder has been moved to
`routes/protected`, and all relevant imports in .tsx files have been updated.

No functional changes are introduced; this is strictly a restructuring effort to
improve maintainability and reduce complexity in the routing module.